+++ includes/lightning.h
@@ -32,9 +32,9 @@ typedef	struct			s_scene_direction
 	t_vector			*direction;
 }						t_scene_direction;
 
-float					calcul_shiny(t_lightning_vectors *l_vectors,
-t_light *light, float length_v, t_vector *vec_l);
-float					compute_special_lights(t_lightning_vectors *l_vectors,
+double					calcul_shiny(t_lightning_vectors *l_vectors,
+t_light *light, double length_v, t_vector *vec_l);
+double					compute_special_lights(t_lightning_vectors *l_vectors,
 t_light *light, t_scene *scene, t_lstobjects *objects);
 t_vector				*type_light(t_lightning_vectors *l_vectors,
 t_light *light, t_scene *scene);
@@ -44,10 +44,10 @@ t_vector				*set_shiny_and_get_color(t_lightning_vectors *l_vectors,
 t_lstobjects *objects);
 int						compute_lightning(t_lightning_vectors *l_vectors,
 t_lstobjects *lights, t_scene *scene, t_lstobjects *objects);
-void					set_color_special_lights(float *intensity,
+void					set_color_special_lights(double *intensity,
 t_l_o_v_s *lovs, t_lstobjects *objects, t_lstobjects *lights);
 void					set_color_ambient(t_vector *color[4],
-float *intensity, t_lstobjects *lights);
+double *intensity, t_lstobjects *lights);
 void					check_light_intersect_sphere(
 t_lightning_vectors *l_vectors, void *object);
 void					check_light_intersect_plan(
@@ -65,8 +65,8 @@ t_scene_direction *s_d, t_light *light);
 void					check_light_intersect_cylinder(
 t_lightning_vectors *l_vectors, void *object, t_scene_direction *s_d,
 t_light *light);
-float					calcul_shiny(t_lightning_vectors *l_vectors,
-t_light *light, float length_v, t_vector *vec_l);
+double					calcul_shiny(t_lightning_vectors *l_vectors,
+t_light *light, double length_v, t_vector *vec_l);
 t_vector				*set_shiny_and_get_color(t_lightning_vectors *l_vectors,
 t_lstobjects *objects);
